+++ libraries/db/include/graphene/db/object.hpp
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 #pragma once
-#include <graphene/db/object_id.hpp>
+#include <graphene/protocol/object_id.hpp>
 #include <fc/io/raw.hpp>
 #include <fc/crypto/city.hpp>
 #include <fc/uint128.hpp>
